+++ test/foundry/BullaFrendLend/BullaFrendLendERC165.t.sol
@@ -14,7 +14,7 @@ contract BullaFrendLendERC165Test is Test {
 
     address admin = makeAddr("admin");
     uint256 constant FEE = 0.01 ether;
-    uint256 constant PROTOCOL_FEE_BPS = 1000; // 10%
+    uint16 constant PROTOCOL_FEE_BPS = 1000; // 10%
 
     function setUp() public {
         bullaClaim = (new Deployer()).deploy_test(address(this), LockState.Unlocked, FEE);
